requirements.txt
einops==0.8.1
joblib==1.4.2
kornia==0.7.3
loguru==0.7.3
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.12.0.88
osgeo==0.0.1
Pillow==12.1.0
pytorch_lightning==1.8.6
rasterio==1.3.11
scikit_learn==1.3.2
scipy==1.16.3
superglue==0.20.0
thop==0.1.1.post2209072238
torch==1.13.1+cu117
torchvision==0.14.1+cu117
yacs==0.1.8